rev: organise navigation strings andreorgnise drawer into componenets
