--- core_ui/src/main/java/com/amalitech/core_ui/navigation/MainNavigationGraph.kt
@@ -1,6 +1,0 @@
-package com.amalitech.core_ui.navigation
-
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.navigation.NavGraphBuilder
-import androidx.navigation.compose.navigation
-
